Spellcircles

Documentation for Spellcircles.

Spellcircles.blank

Spellcircles.circle

Spellcircles.construct

Spellcircles.lines

Spellcircles.polygon

Spellcircles.polygonpt

Spellcircles.polygram

Spellcircles.blank — Method

blank()

Creates a blank spellcircle.

Spellcircles.circle — Method

circle(r)

Draws a circle with radius r.

Arguments

x::Number: Moves polygon by x.

y::Number: Moves polygon by y.

stroke::String: Sets the colour of the lines.

width::Number: Sets the width of the lines.

fill::String: Sets background colour of the polygon.

Spellcircles.construct — Method

construct()

Constructs the spellcircle.

Arguments

name::String: Names the svg file.

size::Int: Size of a square svg from the middle.

pagedimensions::Tuple{Int64, Int64}: Sets of the size of the svg from the middle in the (x,y)-directions.

custom::Bool: If true then the outer circle will not be drawn.

Spellcircles.lines — Method

lines(pt1,pt2)

Draws a line from pt1 to pt2, where each point is a tuple of (x,y).

Arguments

stroke::String: Sets the colour of the lines.

width::Number: Sets the width of the lines.

fill::String: Sets background colour of the polygon.

Spellcircles.polygon — Method

polygon(n,r)

Constructs a regular n-polygon with radius r.

Arguments

orientation::Number: Rotates the polygon clockwise by θ degrees.

x::Number: Moves polygon by x.

y::Number: Moves polygon by y.

stroke::String: Sets the colour of the lines.

width::Number: Sets the width of the lines.

fill::String: Sets background colour of the polygon.

Spellcircles.polygonpt — Method

polygonpt(n,r)

Compute all points of any regular n-polygon with radius r.

The first point is located at (0,r) and all further points are added clockwise.

Arguements

orientation::Number: Rotates the polygon clockwise by θ degrees.

x::Number: Moves polygon by x.

y::Number: Moves polygon by y.

Examples

julia> polygonpt(3,1)
3-element Vector{Tuple{Float64, Float64}}:
  (0.0, 1.0)
  (0.87, -0.5)
  (-0.87, -0.5)

Spellcircles.polygram — Method

polygram(n,r)

Constructs a n-polygram with radius r.

Arguments

orientation::Number: Rotates the polygon clockwise by θ degrees.

x::Number: Moves polygon by x.

y::Number: Moves polygon by y.

stroke::String: Sets the colour of the lines.

width::Number: Sets the width of the lines.

fill::String: Sets background colour of the polygon.

edges::Int: Unfinished.